+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ v2.0.0
 
 Unreleased
 ------------------
+
+11.4.0 - 2019-09-20
+-------------------
 Added
  - Added `__array_interface__` to the Image class.
  - Added `Console.draw_semigraphics` as a replacement for blit_2x functions.